+++ Changelog.md
@@ -1,3 +1,6 @@
+## 0.0.102
+*INCOMPLETE*
+
 ## 0.0.101
 + Added `beta.SGLangInferenceMetadata` metadata type to compile large language models for inference with SGLang.
 + Added `beta.vLLMInferenceMetadata` metadata type to compile large language models for inference with vLLM.